Expose outcome execution boundary in handoffs
