+++ docs/rfcs/RFC-0042-post-trade-outcome-feedback-loop.md
@@ -548,9 +548,10 @@ Acceptance:
 2. source-owner contract tests exist for any upstream change,
 3. manage does not duplicate source-owner calculations,
 4. `EXECUTION_EVIDENCE_BLOCKED` is emitted where execution truth is unavailable, and
-   supportability emits `DPM_OUTCOME_EXTERNAL_EXECUTION_BOUNDARY` with blocked capabilities,
-   required future execution/OMS owner, required `ExternalOrderExecutionAcknowledgement:v1`,
-   acknowledgement-count posture, and deterministic content hash.
+   supportability plus report/AI handoffs emit `DPM_OUTCOME_EXTERNAL_EXECUTION_BOUNDARY` with
+   blocked capabilities, required future execution/OMS owner, required
+   `ExternalOrderExecutionAcknowledgement:v1`, acknowledgement-count posture, and deterministic
+   content hash.
 
 Slice 5 evidence:
 
@@ -808,8 +809,8 @@ fairness analytics, and richer source-owner methodology remain outside the suppo
 | Expected-versus-realized variance decomposition | Supported for source-backed supplied dimensions | Supported where expected and realized evidence reconcile with lineage, hashes, and tests. |
 | Source-degraded outcome review | Supported as explicit state behavior | Degraded, blocked, unsupported, stale, partial, malformed, conflicting, and execution-evidence-blocked states are tested and documented. |
 | Searchable outcome memory | Supported as manage backend memory | Immutable persistence, events, source hashes, run lookup, wave lookup, and indexed search are proven. |
-| Report input from outcome review | Supported as manage handoff contract plus first-wave materialization path | Manage emits bounded report input; `lotus-report`, `lotus-render`, and `lotus-archive` own generated report and archive lifecycle without recomputing outcome truth. |
-| AI evidence input from outcome review | Supported as bounded evidence contract plus governed narrative path | Manage emits bounded AI evidence input; `lotus-ai` owns guarded workflow-pack narrative execution. No recommendation, PM scoring, or client-contact support is claimed. |
+| Report input from outcome review | Supported as manage handoff contract plus first-wave materialization path | Manage emits bounded report input with structured `DPM_OUTCOME_EXTERNAL_EXECUTION_BOUNDARY` evidence; `lotus-report`, `lotus-render`, and `lotus-archive` own generated report and archive lifecycle without recomputing outcome truth or inferring OMS execution. |
+| AI evidence input from outcome review | Supported as bounded evidence contract plus governed narrative path | Manage emits bounded AI evidence input with structured `DPM_OUTCOME_EXTERNAL_EXECUTION_BOUNDARY` evidence; `lotus-ai` owns guarded workflow-pack narrative execution. No recommendation, PM scoring, OMS execution, or client-contact support is claimed. |
 | Gateway outcome composition | Supported through `lotus-gateway` PR #186/#187/#188/#189 | Gateway composes manage outcome-review truth and report/AI handoff posture without recomputing expected values, realized values, variance, tolerance, lineage, freshness, or review state. |
 | Workbench outcome review UX | Supported through `lotus-workbench` PR #146/#147/#148 | Workbench consumes Gateway/BFF contracts only and renders the first-wave outcome-review, report-request, and AI-narrative request posture with canonical proof. |
 | PM operating quality policy, score-run, and fairness-analysis lifecycle | Supported as separate Manage-owned first-wave product | `PUT /api/v1/rebalance/pm-operating-quality/policies/{policy_id}/versions/{policy_version}`, `GET /policies`, and `GET /policies/{policy_id}/versions/{policy_version}` administer immutable policy versions; `POST /score-runs/preview` emits `PmOperatingQualityScoreRun:v1`; `POST /score-runs`, `GET /score-runs`, and `GET /score-runs/{score_run_id}` persist and retrieve immutable score-run evidence; `POST /fairness-analyses/preview` emits bounded `PmOperatingQualityFairnessAnalysis:v1` over persisted score runs and source-defined segments; `POST /fairness-analyses`, `GET /fairness-analyses`, and `GET /fairness-analyses/{fairness_analysis_id}` persist, list, and retrieve immutable fairness-analysis evidence without recomputing score runs. Enabled policies require bank approval and fairness-review evidence; score-run preview/create emits `governance_evidence` and fails closed for missing approval, invalid/expired expiry, unauthorized actors, and missing scoring evidence. Optional `pm_book_scope` materializes source-owned lotus-core `PortfolioManagerBookMembership:v1` scope evidence and fails closed for unavailable, incomplete, degraded, or empty PM-book membership. Optional `peer_group_policy` and `lookback_window_policy` materialize into score-run `scope_evidence`, preserve peer-group/lookback source refs in the content hash, and fail closed with `PM_QUALITY_EVIDENCE_OUTSIDE_LOOKBACK_WINDOW` when dated evidence falls outside the approved window. Fairness analysis validates common policy/as-of scope, requires minimum scorable segment counts, flags governed average-score spread review, persists content-addressed evidence immutably, and does not infer protected classes or rank PMs. Scoring is disabled by default, and HR, compensation, conduct-enforcement, autonomous-ranking, AI-generated scoring, local peer discovery, and source-owner methodology claims are prohibited. |
