+++ src/core/outcomes/__init__.py
@@ -41,6 +41,7 @@ from src.core.outcomes.snapshots import (
     assemble_expected_outcome_snapshot,
 )
 from src.core.outcomes.realized_sources import assemble_realized_outcome_snapshot
+from src.core.outcomes.execution_boundary import build_outcome_external_execution_boundary
 from src.core.outcomes.performance_sources import (
     PerformanceOutcomeSourceError,
     realized_active_performance_source_from_workspace_summary,
@@ -107,6 +108,7 @@ __all__ = [
     "assemble_realized_outcome_snapshot",
     "assert_no_ai_forbidden_fields",
     "build_ai_evidence_input",
+    "build_outcome_external_execution_boundary",
     "build_report_input",
     "compare_outcome_dimension",
     "compare_outcome_dimensions",
