+++ src/core/outcomes/handoffs.py
@@ -8,8 +8,10 @@ from pydantic import BaseModel, Field
 
 from src.core.common.canonical import hash_canonical_payload, strip_keys
 from src.core.portfolio_memory.handoffs import DpmPortfolioMemoryReportContext
+from src.core.outcomes.execution_boundary import build_outcome_external_execution_boundary
 from src.core.outcomes.models import (
     DpmOutcomeDimensionResult,
+    DpmOutcomeExternalExecutionBoundaryEvidence,
     DpmOutcomeSourceRef,
     DpmOutcomeSupportability,
     DpmPostTradeOutcomeReview,
@@ -79,6 +81,12 @@ class DpmOutcomeReportInput(BaseModel):
     section_hashes: dict[str, str] = Field(
         description="Proof-pack section hashes carried by the review."
     )
+    external_execution_boundary: DpmOutcomeExternalExecutionBoundaryEvidence = Field(
+        description=(
+            "Fail-closed external execution/OMS boundary evidence carried for downstream report "
+            "consumers without promoting acknowledgement, fill, settlement, or best-execution truth."
+        )
+    )
     portfolio_memory_context: DpmPortfolioMemoryReportContext | None = Field(
         default=None,
         description=(
@@ -123,6 +131,12 @@ class DpmOutcomeAiEvidenceInput(BaseModel):
         description="AI-safe dimension evidence."
     )
     source_refs: list[DpmOutcomeSourceRef] = Field(description="AI-safe source references.")
+    external_execution_boundary: DpmOutcomeExternalExecutionBoundaryEvidence = Field(
+        description=(
+            "Fail-closed external execution/OMS boundary evidence carried for downstream AI "
+            "consumers without permitting order, fill, settlement, or best-execution claims."
+        )
+    )
     evidence_ref: DpmOutcomeSourceRef = Field(description="Evidence reference for this input.")
     content_hash: str = Field(description="Canonical AI-evidence input hash.")
 
@@ -159,6 +173,7 @@ def build_report_input(
         source_lineage=review.source_lineage,
         source_hashes=review.source_hashes,
         section_hashes=review.section_hashes,
+        external_execution_boundary=build_outcome_external_execution_boundary(review),
         portfolio_memory_context=portfolio_memory_context,
         redaction_policy="NO_RAW_PAYLOADS",
         evidence_ref=_handoff_ref(
@@ -207,6 +222,7 @@ def build_ai_evidence_input(review: DpmPostTradeOutcomeReview) -> DpmOutcomeAiEv
         reason_codes=review.supportability.reason_codes,
         dimensions=dimensions,
         source_refs=_dedupe_source_refs(review),
+        external_execution_boundary=build_outcome_external_execution_boundary(review),
         evidence_ref=_handoff_ref(
             ref_type=OUTCOME_AI_EVIDENCE_REF_TYPE,
             outcome_review_id=review.outcome_review_id,
